+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 ### Fixed
 - Fixes a bug in `sensuctl cluster health` so the correct error is handled.
 - Fixed a bug where assets could not extract git tarballs.
+- Fixed a bug where a check with multiple assets will only install the first asset.
 
 ## [2.0.0-beta.4] - 2018-08-14